+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0]
+
 ## Changed
 
 - hnefatafl-client: don't shutdown TCP on redox.
@@ -237,7 +239,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - hnefatafl-client: Make users and games scrollable.
 - Make all the features dependencies.
 
-[unreleased]: https://github.com/dcampbell24/hnefatafl-copenhagen/compare/v0.13.4...main
+[unreleased]: https://github.com/dcampbell24/hnefatafl-copenhagen/compare/v1.0.0...main
+[1.0.0]:  https://github.com/dcampbell24/hnefatafl-copenhagen/compare/v0.13.4...v1.0.0
 [0.13.4]: https://github.com/dcampbell24/hnefatafl-copenhagen/compare/v0.13.3...v0.13.4
 [0.13.3]: https://github.com/dcampbell24/hnefatafl-copenhagen/compare/v0.13.2...v0.13.3
 [0.13.2]: https://github.com/dcampbell24/hnefatafl-copenhagen/compare/v0.13.1...v0.13.2